Split level 3 tests into separate files
